Refactored logic

+++ src/Models/Ingredient.cs
@@ -3,9 +3,9 @@
     public class Ingredient
     {
         public string Name { get; }
-        public double Amount { get; }
+        public decimal Amount { get; }
 
-        public Ingredient(string name, double amount)
+        public Ingredient(string name, decimal amount)
         {
             Name = name;
             Amount = amount;
