+++ src/Processors/IIngredientProcessor.cs
@@ -4,6 +4,6 @@ namespace OrderProcessingApp.Processors
 {
     public interface IIngredientProcessor
     {
-        Dictionary<string, double> CalculateTotalIngredients(IEnumerable<Order> orders, Dictionary<string, Product> productMap, Dictionary<string, List<Ingredient>> ingredientProductMap);
+        Dictionary<string, decimal> CalculateTotalIngredients(IEnumerable<Order> orders, Dictionary<string, Product> productMap, Dictionary<string, List<Ingredient>> ingredientProductMap);
     }
 }
